Reference

Charon has three exported functions, MCMCsampler, exactposterior and unpackposterior, that can be used after typing

using Charon 

Charon.Charon — Module

module Charon

This package provides an MCMC sampler and other functions to infer drift times and the number of individuals from environmental DNA.

MCMCsampler

This is the MCMC sampler. It has seven methods.

Data formats:

You can input the data in four formats.

With two (opened) (gzipped) files: a base count file, which is a CSV file where the first column are the number of derived reads, and the second column is the coverage, and a frequency file, which is a CSV file with one column, containing the frequencies. The data at line i in the base count file corresponds to the same locus as the data on line i of the frequency file.

As an (opened) (gzipped) DICE file, or in the form of a DataFrame, also in DICE format. So the first column is the number of ancestral reads, the second column is the number of derived reads, the third column is the frequency in the anchor population, and the fourth column is the count of the number of loci where this combination of three numbers occur.

Or, by providing three vectors: coverages, derivedreads, frequencies, of length equal to the number of SNPs, where at locus i, there are derivedreads[i] derived reads, coverages[i] coverage and frequencies[i] frequency in the anchor population.

The third format is given with four vectors: coverages, derivedreads, frequencies, counts. This means that there are counts[i] loci with coverages[i] coverage, derivedreads[i] derived reads and frequency frequencies[i] in the anchor population.

If you provide data in formats 1, 2, or 3, then the program will automatically convert it to format 4.

Parameters:

nchains, number of chains, which is a positive integer. If you want to run all your chains in parallel, start julia with number of threads equal to nchains.

nsteps number of samples per chain, which is a positive integer.

prioronn the prior on n, specified as a subtype of DiscreteUnivariateDistribution of the Distributions Julia package. Our implementation requires that prioronn has support on {1, 2, 3, 4, 5, 6, 7, 8, 9, 10}, because otherwise rounding errors will accumulate too much.

prioronτCτA is the prior on (τC, τA), which allows for correlation between τC and τA. Its type is a subtype of ContinuousMultivariateDistribution. It should have support contained in [0,∞)x[0,∞).

prioronϵ is the prior on ϵ. It should have support in [0, 0.5). It is a subtype of ContinuousUnivariateDistribution in the Distributions package.

coverages a vector with coverages = ancestral reads + derived reads. Is a subtype of AbstractVector{<:Integer}. All coverages should be non-negative integers, and at least one should be positive.

derivedreads a vector of derived reads. Is a subtype of AbstractVector{<:Integer}. All elements of the vector should be non-negative integers.

frequencies a vector of frequencies. Is a subtype of AbstractVector{<:Real}. Each frequency is between 0.0 and 1.0. At least one frequency should be strictly between 0.0 and 1.0 with corresponding positive coverage.

counts, all elements should be non-negative. For each index, counts[index] indicates how many loci there are with derivedreads[index] derived reads, coverage coverages[index] and frequency frequencies[index].

df a DataFrame from the DataFrames package in the DICE-2 format. So the first column should be the number of ancestral reads, the second column the number of derived reads, the third column the frequencies in the anchor population, and the fourth column the counts of how many times this particular combination of ancestral reads, derived reads and frequency occurs.

dicefile is either an opened (gzipped) DICE file, or a path to a (gzipped) DICE file.

Keyword parameters. Keyword parameters should be given as keyword=value to the function, in case you want to set another value then the default.

messages is an integer. If messages is non-positive, no message will be printed. If messages is positive, every messages steps a message will be printed with the progress of the sampler. The default value is nsteps÷100, so every 1% progress a message is printed.

scalingmessage is either true (default) or false. If true, a message will be printed when the scaling constant changes.

header is nothing, true (default) or false. Has the dicefile a header? If nothing, the software tries to determine whether the dicefile has a header. This works only when you provide a path to a file.

The output is a vector with nchains items. Each item represents a chain. Each item is a tuple consisting of six columns, as described here.

Charon.MCMCsampler — Function

MCMCsampler(nchains::Integer, nsteps::Integer, prioronn::DiscreteUnivariateDistribution, prioronτCτA::ContinuousMultivariateDistribution, prioronϵ::ContinuousUnivariateDistribution, coverages::AbstractVector{<:Integer}, derivedreads::AbstractVector{<:Integer}, frequencies::AbstractVector{<:Real}, counts::AbstractVector{<:Integer}; messages::Integer=nsteps÷100, scalingmessages::Bool=true)

The MCMC sampler as descriped in the article. Parameters:

Number of chains and samples:

nchains a positive integer. Indicates the number of chains. If you start Julia with nchains threads, then the chains are run in parallel.

nsteps a positive number. The number of samples for each chain.

Prior

prioronn the prior on n.

prioronτCτA the prior on τC and τA. The sampler allows correlation in the prior between the parameters.

prioronϵ the prior on ϵ.

Data

coverages a vector of positive integers, the coverages.

derivedreads a vector of non-negative integers, the number of derived reads. for each index 0 ≤ derivedreads[i] ≤ coverages[i].

frequencies a vector of real number in the interval [0, 1]. For at least one index i, coverages[i]>0 and 0 < frequencies[i] < 1.

counts a vector with positive integers. For each index counts[i] indicates in how many loci there is coverage coverages[i], derivedreads[i] derived reads and frequency frequencies[i].

Keyword parameters

messages an integer. If the integer is non-positive, no messages will be printed. If messages is positive, every messages steps, an update about the progress will be printed. The default value is nsteps÷100, so every 1% progress a message is printed.

scalingmessages should the sampler print a message when the scaling constants change? The default value is true.

Output: The output is a vector of tuples. Each tuple represents a chain. Each tuple has the following elements: nsample, τCsample, τAsample, ϵsample, accepted, logjointprob.

nsample sample for the number of individuals.

τCsample a nsample x maximum(prioronn) matrix, where the n-th column is the MCMC chain of τC conditioned on n individuals.

τAsample same as τCsample but now for τA.

ϵsample same as τCsample but now for ϵ.

accepted a nsample x (maximum(prior)+1) matrix, where the first column indicates whether the proposal for n was accepted. The n+1-th column indicates whether the proposal for (τC,τA,ϵ) was accepted for the posterior conditioned on n.

logjointprob a nsample x (maximum(prior)+1) matrix, where the first column is the log joint probability, up to a constant not depending on parameters, of the unconditioned posterior, and the n+1-th column is the log joint probability, up to a constant not depending on the parameters, of the posterior conditioned on n individuals.

MCMCsampler(nchains::Integer, nsteps::Integer, prioronn::DiscreteUnivariateDistribution, prioronτCτA::ContinuousMultivariateDistribution, prioronϵ::ContinuousUnivariateDistribution, coverages::AbstractVector{<:Integer}, derivedreads::AbstractVector{<:Integer}, frequencies::AbstractVector{<:Real}; messages::Integer=nsteps÷100, scalingmessages::Bool=true)

Same as the other method, except with data coverages, derivedreads and frequencies. This is the data of all loci, so at locus i, the coverage is coverages[i] with derivedreads[i] derived reads, and frequency frequencies[i] in the anchor population.

MCMCsampler(nchains::Integer, nsteps::Integer, prioronn::DiscreteUnivariateDistribution, prioronτCτA::ContinuousMultivariateDistribution, prioronϵ::ContinuousUnivariateDistribution, df::AbstractDataFrame; messages::Integer=nsteps÷100, scalingmessages::Bool=true)

Same as the other method, but now with data df in the DICE 2 format: It should contain exactly four columns, the first one contains the number of ancestral reads, the second the number of derived reads, the third the frequency of the derived allele in the anchor population and the fourth the number of loci with exactly this combination of data.

MCMCsampler(nchains::Integer, nsteps::Integer, prioronn::DiscreteUnivariateDistribution, prioronτCτA::ContinuousMultivariateDistribution, prioronϵ::ContinuousUnivariateDistribution, dicefile::IO; messages::Integer=nsteps÷100, scalingmessages::Bool=true, header::Bool=true)

Same as the other method, but with dicefile an opened CSV-file in DICE 2 format, as decribed above.

If you have Julia 1.3 or higher, and the file is a zipped file, open it using a zip decompressor:

using CodecZlib
dicefile = GzipDecompressorStream(open("path/to/your/file.csv.gz"))

With the named argument "header" one should indicate whether the CSV file has a header. The default is true.

MCMCsampler(nchains::Integer, nsteps::Integer, prioronn::DiscreteUnivariateDistribution, prioronτCτA::ContinuousMultivariateDistribution, prioronϵ::ContinuousUnivariateDistribution, dicefile::AbstractString; messages::Integer=nsteps÷100, scalingmessages::Bool=true, header::Union{Nothing, Bool}=nothing)

Same as the other methods, except that dicefile is a path to a (gzipped) CSV-file in DICE 2 format, as decribed above.

WARNING: If you work with gzipped files, Julia 1.3 or higher is required.

MCMCsampler(nchains::Integer, nsteps::Integer, prioronn::DiscreteUnivariateDistribution, prioronτCτA::ContinuousMultivariateDistribution, prioronϵ::ContinuousUnivariateDistribution, sedimentdata::Union{AbstractString, IO}, frequencies::Union{AbstractString, IO}; messages::Integer=nsteps÷100, scalingmessages::Bool=true, headersedimentdata::Union{Nothing, Bool}=nothing, headerfrequencies::Union{Nothing, Bool}=nothing)

sedimentdata is a path to a (gzipped) CSV file, or an opened (gzipped) file, where the first column are the derived reads, and the second column are the coverages. frequencies is a path to a (gzipped) CSV file with one column, or an opened (gzipped) file with frequencies, containing frequencies of the anchor population. The number of rows of both files should be the same.

MCMCsampler(nchains::Integer, nsteps::Integer, prioronn::DiscreteUnivariateDistribution, prioronτCτA::ContinuousMultivariateDistribution, prioronϵ::ContinuousUnivariateDistribution, sedimentdata::DataFrame, frequencies::DataFrame; messages::Integer=nsteps÷100, scalingmessages::Bool=true)

sedimentdata is a path to a DataFrame, where the first column are the derived reads, and the second column are the coverages. frequencies is a DataFrame with one column, containing frequencies of the anchor population. The number of rows of both files should be the same.

unpackposterior

This function is used to build the unconditional posterior from the MCMC samples conditioned on n, as described in the paper. It has two methods.

Arguments:

nsteps is the number of MCMC samples, a positive integer.

chains this the tuple that is the output of MCMCsampler.

Charon.unpackposterior — Function

unpackposterior(nsteps::Integer, chains::AbstractVector{<:Tuple{AbstractVector{<:Integer}, AbstractMatrix{<:Real}, AbstractMatrix{<:Real}, AbstractMatrix{<:Real}, AbstractMatrix{Bool}, AbstractMatrix{<:Real}}})

Calculate the unconditional posterior from the output of MCMCsampler. chains is the output of the sampler, nsteps is the number of MCMC samples. This function returns a DataFrame containing nsteps rows, and 7 columns:

nsample sample for the number of individuals.

τCsample sample for τC.

τAsample sample for τA.

ϵsample sample for ϵ.

accepted Was any of the conditioned MCMC sample proposals, or the proposal for n accepted at step i?

logjointprob the log joint probability up to a constant not depending on the parameters.

chainid the id of the chain. A number 1,...,nchains.

unpackposterior(chains::AbstractVector{<:Tuple{AbstractVector{<:Integer}, AbstractMatrix{<:Real}, AbstractMatrix{<:Real}, AbstractMatrix{<:Real}, AbstractMatrix{Bool}, AbstractMatrix{<:Real}}})

Calculate the unconditional posterior from the output of MCMCsampler. chains is the output of the sampler. The number of samples is determined from chains. The output is the same as the other method.

exactposterior

exactposterior is a function to calculate the posterior up to a fixed constant, only depending on the data, but not on the parameters. You can use this function for maximum posterior estimation. If you use uniform priors, you can use this function for maximum likelihood estimation. MCMCsampler uses this function to find a good starting point for the sampler. It has two methods.

The posterior is calculated for each combination of parameters (n, τC, τA, ϵ) with n in nrange, τC in τCrange, τA in τArange and ϵ in ϵrange. So make sure that length(nrange)*length(τCrange)*length(τArange)*length(ϵrange) is not too large, as otherwise it will take a very long time and you might run out of memory.

Parameters:

nrange vector of n values. Subtype of AbstractVector{<:Integer}. Should be a subset of {1, 2, 3, 4, 5, 6, 7, 8, 9, 10}.

τCrange vector of τC values. Subtype of AbstractVector{<:Real}. All values should be non-negative.

τArange vector of τA values. Subtype of AbstractVector{<:Real}. All values should be non-negative.

ϵrange vector of ϵ values. Is a subtype of AbstractVector{<:Real}. All values should be non-negative and smaller than 0.5.

coverages a vector with coverages = ancestral reads + derived reads. Is a subtype of AbstractVector{<:Integer}. All coverages should be non-negative integers, and at least one should be positive.

uniquecoverages should be equal to unique(coverages).

derivedreads a vector of derived reads. Is a subtype of AbstractVector{<:Integer}. All elements of the vector should be non-negative integers.

frequencies a vector of frequencies. Is a subtype of AbstractVector{<:Real}. Each frequency is between 0.0 and 1.0. At least one frequency should be strictly between 0.0 and 1.0 with corresponding positive coverage.

counts, all elements should be non-negative. For each index, counts[index] indicates how many loci there are with derivedreads[index] derived reads, coverage coverages[index] and frequency frequencies[index].

Keyword argument.

messages is an integer. If messages is non-positive, no message will be printed. If messages is positive, every messages steps a message will be printed with the progress of the calculations.

The output are 5 vectors: ns, τCs, τAs, ϵs, logliks, of each of length length(n)*length(τCrange)*length(τArange)*length(ϵrange), where logliks[index] is the log likelihood, up to an additive constant, with parameters ns[index], τCs[index], τAs[index] and ϵs[index]. The additive constant only depends on the data, but not on the parameters.

Maximum posterior estimation (MAP estimator)

Suppose you have results

ns, τCs, τAs, ϵs, logliks = exactposterior(args...) 

where args are your arguments (data and priors, etc.). Then you can calculate the index where the log likelihood is maximised:

i_max = argmax(logliks)

So

ns[i_max], τCs[i_max], τAs[i_max], ϵs[i_max]

is the MAP estimator. If you provide uniform priors, then the MAP estimator is a maximum likelihood estimator, where you maximise over the domain of the prior.

Charon.exactposterior — Function

exactposterior(nrange::AbstractVector{<:Integer}, τCrange::AbstractVector{<:Real}, τArange::AbstractVector{<:Real}, ϵrange::AbstractVector{<:Real}, coverages::AbstractVector{<:Integer}, uniquecoverages::AbstractVector{<:Integer}, derivedreads::AbstractVector{<:Integer}, frequencies::AbstractVector{<:Real}, counts::AbstractVector{<:Integer}; messages::Integer=0)

Calculate the posterior for every combination of parameters (n, τC, τA, ϵ) where n is in nrange, τC is in τCrange, τA is in τArange and ϵ is in ϵrange. This function returns vectors ns, τCs, τAs, ϵs, logliks. So the log posterior with ns[i] individuals, drift parameters τCs[i] and τAs[i] and error rates ϵs[i] is logliks[i]. The vectors are of length length(n)*length(τCrange)*length(τArange)*length(ϵrange). Make sure that this is not too large, as otherwise the calculation might take very long time, or you might run out of memory.

coverages, derivedreads, frequencies, and counts are the data. So there are counts[i] loci with coverage coverages[i], derivedreads[i] derived reads and frequency frequencies[i] in the anchor population. uniquecoverages=unique(coverages).

messages is an integer. If it is positive, every message step an update is printed about the progress. If messages is non-positive, no update will be printed.

The output are four vectors ns, τCs, τAs, ϵs, logliks. For each index logliks[i] is the log likelihood with ns[i] individuals, drift parameters τCs[i] and τAs[i] and error rate ϵs[i].

exactposterior(nrange::AbstractVector{<:Integer}, τCrange::AbstractVector{<:Real}, τArange::AbstractVector{<:Real}, ϵrange::AbstractVector{<:Real}, coverages::AbstractVector{<:Integer}, derivedreads::AbstractVector{<:Integer}, frequencies::AbstractVector{<:Real}; messages::Integer=length(n)*length(τCrange)*length(τArange)*length(ϵrange)÷100)

A user-friendly function to calculate the log posterior. The input are data in the form of three vectors: coverages, derivedreads and frequencies. This is the data for all loci. So at locus i, the coverage is coverages[i], the number of derived reads is derivedreads[i] and the frequency in the anchor population is frequencies[i]. The other parameters are the same as in the other method.

The output is the same as the other method. The default value for messages is length(n)*length(τCrange)*length(τArange)*length(ϵrange)÷100), so every 1% progress an update is printed.

Charon EigenExpansion type

This type is not exported, but can be used after using Charon: EigenExpansion, or via Charon.EigenExpansion. It represents a matrix decomposition $M=P*D*P^{-1}$, where $P$ is an invertible matrix with inverse $P^{-1}$, and $D$ is a diagonal matrix. It is used to efficiently calculate the exponent of a matrix.

Charon.EigenExpansion — Type

EigenExpansion{Q<:AbstractMatrix{<:Real}, R<:Real, S<:AbstractVector{R}, T<:AbstractMatrix{<:Real}}

Representation of a matrix decomposition M=P*D*P^{-1}, where D is a diagonal matrix and P an invertible matrix with inverse Pinv.

This decomposition can be used for fast matrix exponentiation, using that exp(M) = P*exp(D)*Pinv, and the fact that the exponent of a diagonal matrix is formed by exponentiating its diagonal elements.

Fields:

P::Q 
D::Diagonal{R, S}
Pinv::T 

Internal, non-exported functions

The following functions are not exported, and only available via Charon.functionname, or using Charon: functionname.

Charon.calcloglik — Function

calcloglik(coverages::AbstractVector{<:Integer}, derivedreads::AbstractVector{<:Integer}, counts::AbstractVector{<:Integer}, M::AbstractMatrix{<:Real}, hvals::AbstractMatrix{<:Real}, q::AbstractDict{<:Tuple{Integer, Integer}, <:AbstractVector{<:Real}}, index::Integer)

Calculate the log-likelihood of the data at index index. Parameters:

coverages vector with coverages.

derivedreads vector with derived reads.

counts counts[i] loci have coverages[i] coverage and derivedreads[i] derived reads.

M=e^{Q*τA}e^{QꜜτC}, where τC and τA are the drift parameters, Q and Qꜜ are the matrices as in Schraiber 2018, and M is calculated with calcmatrix.

hvals the matrix constructed with preparedata, see the documentation of preparedata.

q the dictionary made with makeq.

index the index of the data we are working on.

This function returns a real number.

Charon.calcmatrix — Function

calcmatrix(τC::Real, τA::Real, Qꜜ::EigenExpansion, Q::EigenExpansion)

Calculate the matrix e^{Q*τ_A}e^{Qꜜτ_C} as in Schraiber. Note that in general, for matrices, exp(A+B) ≠ exp(A)exp(B) (equality holds if and only if A and B commute, which is not the case for Q and Qꜜ).

Charon.filtervectorsandapplycountmap — Function

filtervectorsandapplycountmap(coverages::AbstractVector{<:Integer}, derivedreads::AbstractVector{<:Integer}, frequencies::AbstractVector{<:Real}, allowedindices::AbstractVector{Bool})

This function does the following:

coverages_filtered = coverages[allowedindices]
derivedreads_filtered = derivedreads[allowedindices]
frequencies_filtered = frequencies[allowedindices]

Next, it counts how many times (coverages_filtered[i], derivedreads_filtered[i], frequencies_filtered[i]) occurs, for all triples, and next returns four vectors (coverages_filtered_unique, derivedreads_filtered_unique, frequencies_filtered_unique, counts), where each triple (coverages_filtered_unique[i], derivedreads_filtered_unique[i], frequencies_filtered_unique[i]) is unique and occurs counts[i] times in zip(coverages_filtered, derivedreads_filtered, frequencies_filtered).

Charon.logprobderivedreads! — Function

logprobderivedreads!(q::AbstractDict{<:Tuple{Integer, Integer}, <:AbstractVector{<:Real}}, n::Integer, τC::Real, τA::Real, ϵ::Real, coverages::AbstractVector{<:Integer}, uniquecoverages::AbstractVector{<:Integer}, derivedreads::AbstractVector{<:Integer}, counts::AbstractVector{<:Integer}, Qꜜ::EigenExpansion, Q::EigenExpansion, binomialcoefficients::AbstractVector{<:Integer}, hvals::AbstractMatrix{<:Real})

Calculate the log probability of the data, a.k.a. the likelihood. q is mutated with the function updateq!. q is a dictionary, and is used for intermediate calculations.

Parameters:

n number of individuals.

τC and τA drift parameters.

ϵ error rate.

Data:

coverages the coverages.

uniquecoverages=unique(coverages)

derivedreads vector with the derived reads.

counts in counts[i] loci there are coverages[i] coverage and derivedreads[i] derived reads.

Other data:

Qꜜ the matrix Qꜜ as in Schraiber 2018.

Q the matrix Q as in Schraiber 2018.

binomialcoefficients vector of length 2n+1, where entry k+1 is the binomial coefficient of 2n over k. Note that Julia is 1-based.

hvals the matrix as constructed with the function preparedata. See documentation for preparedata.

Charon.makeq — Function

makeq(nmax::Integer, coverages::AbstractVector{<:Integer}, derivedreads::AbstractVector{<:Integer})

Construct a vector of length nmax, where the n-th item is a dictionary with keys (coverages[i], derivedreads[i]) and values real vectors of length 2n+1. The vectors contain arbitrary data.

Charon.makeqfixedn — Function

makeqfixedn(n::Integer, uniquecoveragesandderivedreads::AbstractVector{<:Tuple{Integer, Integer}})

Construct a dictionary with keys uniquecoveragesandderivedreads[i] (which is a tuple where the first element is the coverage and the second is the number of derived reads) and values real vectors of length 2n+1. The values in the vectors are arbitrary.

makeqfixedn(n::Integer, coverages::AbstractVector{<:Integer}, derivedreads::AbstractVector{<:Integer})

Construct a dictionary with keys (coverages[i], derivedreads[i]) and values real vectors of length 2n+1. The values in the vectors are arbitrary.

Charon.makeQ — Function

makeQ(n::Integer)

Calculate the (2n+1)x(2n+1) tridiagonal matrix Q as in Schraiber 2018.

Charon.makeQꜜ — Function

makeQꜜ(n::Integer)

Calculate the (2n+1)x(2n+1) tridiagonal matrix Qꜜ as in Schraiber 2018.

Charon.preparedata — Function

preparedata(n::Integer, frequencies::AbstractVector{<:Real})

Returns a named tuple (Qꜜ=Qꜜ, Q=Q, binomialcoefficients=binomialcoefficients, hvals=hvals), where Qꜜ and Q are EigenExpansions of Qꜜ and Q (as in Schraiber 2018), respectively. hvals is a (2n+1) x length(y) matrix of type SizedMatrix. SizedMatrix is used rather than SMatrix because this matrix is so large and SMatrix is slow for large matrices. The matrix hvals is defined as hvals[k+1,i] = frequencies[i]^k*(1-frequencies[i])^(2n-k) (note that Julia has 1-based indexing). binomialcoefficients is a vector of length 2n+1 (of type SVector) containing the binomial coefficients 2n over k, k=0,...,2n.

Charon.readcsvfile — Function

readcsvfile(filename::AbstractString, header::Union{Nothing, Bool})

Read CSV file and turn it into a DataFrame (from DataFrames package). If header=nothing, it automatically detects whether the CSV file has a header.

WARNING: this function assumes that the data in the first column does not contain Strings, otherwise it might wrongly infer that the file has a header, so you will miss one row of data.

readcsvfile(filename::IO, header::Bool)

Read the opened CSV file. header indicates whether the file has a header.

Charon.updateq! — Function

updateq!(q::AbstractDict{<:Tuple{Integer, Integer}, <:AbstractVector{<:Real}}, n::Integer, ϵ::Real, uniquecoverages::AbstractVector{<:Integer}, binomialcoefficients::AbstractVector{<:Integer})

Update the dictionary q, so thatq[(R, d)][k+1]is the probability of d derived reads out of R when you have k derived alleles withninidividuals and error rateϵ.uniquecoveragesare all unique occurences of coverages in the data.binomialcoefficients` is a vector of length 2n+1, with k+1-th entry equal to the binomial coefficients of 2n over k. Note that Julia has 1-based indexing.

Extended Julia base functions.

I extended several Julia base functions to my custom EigenExpansion type.

Base.exp — Function

exp(s::Real, F::EigenExpansion)

Calculate e^(s*M) in a numerically efficient way.

If F=EigenExpansion(M), then exp(s*M)=exp(s, F)=F.P*exp(s*F.D)*F.Pinv, which makes use of the fast implementation of exp for diagonal matrices (it is just the exponentiation of the diagonal elements).

Base.show — Function

Base.show(io::IO, mime::MIME{Symbol("text/plain")}, F::EigenExpansion{Q, R, S, T}) where {Q<:AbstractMatrix{<:Real}, R<:Real, S<:AbstractVector{R}, T<:AbstractMatrix{<:Real}}

Pretty display of EigenExpansion types.

Base.:== — Function

==(E1::EigenExpansion, E2::EigenExpansion)

Test whether E1 and E2 have the same decomposition P*D*P^(-1), so whether E1.P == E2.P, E1.D == E2.D and E1.Pinv == E2.Pinv.